requirements.txt
certifi==2023.5.7
charset-normalizer==3.1.0
idna==3.4
pybit==5.3.0
python-dotenv==1.0.0
requests==2.31.0
urllib3==2.0.2
websocket-client==1.5.2
websockets==11.0.3
python-dotenv
